% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFA-functions.R
\name{cfa_string_builder}
\alias{cfa_string_builder}
\title{Build a string for CFA in lavaan based on item names}
\usage{
cfa_string_builder(dat, drop.items = NULL)
}
\arguments{
\item{dat}{The raw data so that kperscale and scale.names can be calculated instead of specified. NOT YET IMPLEMENTED.}

\item{drop.items}{A list of vectors of item numbers to be dropped WITHIN each scale. These are the column numbers within each scale's sub-data.frame, not the overall column numbers.}
}
\description{
Build a string for CFA in lavaan based on item names
}
