% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRT-functions.R
\name{comparer}
\alias{comparer}
\title{Compare PCM, GPCM, and GRM models}
\usage{
comparer(
  dat,
  scale.names = NULL,
  drop.items = NULL,
  pval = 0.01,
  verbose = TRUE
)
}
\arguments{
\item{dat}{}

\item{scale.names}{}

\item{drop.items}{}

\item{pval}{}

\item{verbose}{}
}
\value{

}
\description{
Compare PCM, GPCM, and GRM models
}
