% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EFA-functions.R
\name{create_links_EFA}
\alias{create_links_EFA}
\title{Create Links for Sankey Diagram from EFA Loadings}
\usage{
create_links_EFA(
  loadings,
  cutoff = 0.4,
  scale_names = NULL,
  factor_names = NULL,
  item_suffix_regex = NULL
)
}
\arguments{
\item{loadings}{Factor loadings from EFA}

\item{cutoff}{Minimum loading value used to determine if an item loads on a factor}

\item{scale_names}{Vector of scale names; optional (must be based on common part of item names)}

\item{factor_names}{Vector of factor names; optional}

\item{item_suffix_regex}{String; a regular expression passed to identify the part of the item names that is NOT part of the construct name. This is used to dynamically generate construct names based on the provided data with the default for the MASDER style names. That is, in the item acad_sc_10, this regular expression matches _10. As an alternative, scale_names can be specified so regular expressions can be avoided.}
}
\description{
Create Links for Sankey Diagram from EFA Loadings
}
