% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRT-functions.R
\name{grmit}
\alias{grmit}
\title{Runs the GRM model for each scale}
\usage{
grmit(dat, scale.names = NULL, drop.items = NULL, verbose = TRUE)
}
\arguments{
\item{dat}{}

\item{scale.names}{}

\item{drop.items}{}

\item{verbose}{}
}
\value{

}
\description{
Runs the GRM model for each scale
}
