% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRT-functions.R
\name{itemfitbuilder}
\alias{itemfitbuilder}
\title{Build a table of item fit statistics.}
\usage{
itemfitbuilder(out)
}
\arguments{
\item{out}{A list of mirt output objects. This function will loop over this list.}
}
\description{
Build a table of item fit statistics.
}
