% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRT-functions.R
\name{itemplotter}
\alias{itemplotter}
\title{Generate the item plots from grmit}
\usage{
itemplotter(out, guesspar = TRUE)
}
\arguments{
\item{out}{}

\item{guesspar}{}
}
\value{

}
\description{
Generate the item plots from grmit
}
