% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-string-functions.R
\name{model_string_builder}
\alias{model_string_builder}
\title{Create Model Strings for Analysis Functions Based on List of Items}
\usage{
model_string_builder(
  dat = NULL,
  scale_names = NULL,
  kperscale = NULL,
  drop_items = NULL,
  keep_items = NULL,
  str_type = c("cfa", "mirt", "bfactor"),
  combo_scales = NULL,
  simplify = TRUE,
  item_suffix_regex = "_([^_]*)$"
)
}
\arguments{
\item{dat}{Dataset to be analyzed. If provided, scale_names and kpersecale will be computed from this (assumes data uses internal MASDER naming convention). This is necessary if str_type is "mirt" or "bfactor".}

\item{scale_names}{A vector of scale names. Determined from dat if provided.}

\item{kperscale}{A table of the number of items per scale. Determined from dat if provided.}

\item{drop_items}{A list of vectors of item numbers to drop for each scale; named components are okay. Any scales that will not have any items dropped should be indicated with NULL instead of a vector of item numbers to drop.}

\item{keep_items}{A list of vectors of item numbers to keep for each scale; named components are okay. This format is preferred to drop_items. If no items to drop or keep are provided, all items in dat will be kept.}

\item{str_type}{A string indicating what type of analysis the function should build a string for: "cfa" indicates a simple factor structure for use in lavaan, "mirt" indicates a correlated traits model for use with the mirt::mirt.model function, and "bfactor" indicates a bifactor model for use with mirt::bfactor.}

\item{combo_scales}{A list of additional scales and items. This is needed if scales will be created that do not use the typical naming convention.}

\item{simplify}{Logical; if TRUE and str_type is of length 1, an outer list will not be returned, just the value of the component of the list.}

\item{item_suffix_regex}{String; a regular expression passed to identify the part of the item names that is NOT part of the construct name. This is used to dynamically generate construct names based on the provided data with the default for the MASDER style names. That is, in the item acad_sc_10, this regular expression matches _10. As an alternative, scale_names can be specified so regular expressions can be avoided.}
}
\value{
List with named components if simplify = FALSE or str_type has length 2 or more.
}
\description{
Create Model Strings for Analysis Functions Based on List of Items
}
