% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRT-functions.R
\name{princaller}
\alias{princaller}
\title{Runs princals and can compare linear vs. ordinal}
\usage{
princaller(
  dat,
  scale.names = NULL,
  items = NULL,
  method = "both",
  verbose = FALSE,
  plot_both = FALSE,
  return_objs = TRUE
)
}
\arguments{
\item{dat}{Data from items; usually a data.frame.}

\item{scale.names}{Vector of scale names; must be based on common part of item names.}

\item{items}{A vector of column numbers (corresponding to columns in dat) to be plotted. Rarely used - usually leave as NULL.}

\item{method}{If method = "both" the output is a list containing the first two eigenvalues and the loss function; if it "linear" or "ordinal" the primary output is loadings plots}

\item{verbose}{Logical for debug purposes.}

\item{plot_both}{Logical; If method = "both" should the loadings plots be created?}
}
\value{

}
\description{
Runs princals and can compare linear vs. ordinal
}
