% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRT-functions.R
\name{itemplotter}
\alias{itemplotter}
\title{Generate the item plots from grmit}
\usage{
itemplotter(out, guesspar = FALSE, layoutcol = NULL, ...)
}
\arguments{
\item{out}{A list of objects created by mirt (one object per scale). This function loops over this list.}

\item{guesspar}{Old attempt at creating a grid array. NOT WORKING.}

\item{layoutcol}{Integer number of figures to print per row (number of columns in a grid of images). This does not work well; better to use knitr options fig.show = "hold" and out.width = "50\%" (with appropriate adjustments, e.g., 3 figures per row is 33\%).}

\item{...}{Options to be passed to mirt::itemplot}
}
\description{
Generate the item plots from grmit
}
