% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EFA-functions.R
\name{make_sankey_EFA}
\alias{make_sankey_EFA}
\title{Title}
\usage{
make_sankey_EFA(loadings, ...)
}
\arguments{
\item{loadings}{Factor loadings from EFA}

\item{...}{Options to be passed to create_links_EFA}
}
\value{

}
\description{
Title
}
