% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRT-functions.R
\name{princaller}
\alias{princaller}
\title{Runs princals and can compare linear vs. ordinal}
\usage{
princaller(
  dat,
  scale.names = NULL,
  items = NULL,
  method = "both",
  verbose = FALSE,
  plot_both = FALSE
)
}
\arguments{
\item{dat}{}

\item{scale.names}{}

\item{items}{}

\item{method}{If method = "both" the output is a list containing the first two eigenvalues and the loss function; if it "linear" or "ordinal" the primary output is loadings plots}

\item{verbose}{}

\item{plot_both}{If method = "both" should the loadings plots be created?}
}
\value{

}
\description{
Runs princals and can compare linear vs. ordinal
}
